requirements.txt
attrs==25.4.0
aws-cdk-lib==2.232.1
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==48.20.0
black==25.12.0
boto3==1.42.18
botocore==1.42.18
cattrs==25.3.0
click==8.3.1
constructs==10.4.4
importlib_resources==6.5.2
iniconfig==2.3.0
jmespath==1.0.1
jsii==1.122.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
publication==0.0.3
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
ruff==0.14.10
s3transfer==0.16.0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.6.2
